+++ backend/MobyPark-api/Dtos/ParkingLot/ParkingLotDto.cs
@@ -9,8 +9,8 @@ namespace MobyPark_api.Dtos.ParkingLot
         public string Location { get; set; } = string.Empty;
         public string Address { get; set; } = string.Empty;
         public long Capacity { get; set; }
-        public float? Tariff { get; set; }
-        public float? DayTariff { get; set; }
+        public decimal? Tariff { get; set; }
+        public decimal? DayTariff { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public string Coordinates { get; set; }
     }
